requirements_qafe.txt
allennlp==1.1.0
beautifulsoup4==4.13.3
blis==0.4.1
boto3==1.36.25
botocore==1.36.25
catalogue==1.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==7.1.2
cymem==2.0.11
edlib==1.3.9.post1
exceptiongroup==1.2.2
filelock==3.0.12
fsspec==2025.2.0
future==1.0.0
gdown==5.2.0
h5py==3.11.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
jmespath==1.0.1
joblib==1.4.2
jsonnet==0.20.0
jsonpickle==4.0.2
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.1
nltk==3.9.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
overrides==3.1.0
packaging==24.2
plac==1.1.3
pluggy==1.5.0
preshed==3.0.9
protobuf==5.29.3
PySocks==1.7.1
pytest==8.3.4
python-dateutil==2.9.0.post0
qaeval==0.1.0
regex==2024.11.6
requests==2.32.3
s3transfer==0.11.2
sacremoses==0.1.1
scikit-learn==1.3.2
scipy==1.10.1
sentencepiece==0.2.0
six==1.17.0
soupsieve==2.6
spacy==2.2.4
srsly==1.0.7
sympy==1.13.3
tensorboardX==2.6.2.2
thinc==7.4.0
threadpoolctl==3.5.0
tokenizers==0.8.1rc1
tomli==2.2.1
torch==2.4.1
tqdm==4.67.1
transformers==3.0.2
triton==3.0.0
typing_extensions==4.12.2
urllib3==1.26.20
wasabi==0.10.1
